requirements.txt
asgiref==3.9.0
Django==5.2.4
django-phonenumber-field==8.1.0
gunicorn==23.0.0
packaging==25.0
phonenumbers==9.0.8
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.9.0
